+++ NEWS.md
@@ -4,6 +4,7 @@ pekwm-0.4.2
 Closed issues
 -------------
 
+* #205, SysTray icons not displaying properly due to missing configure notify.
 * Fix out of bounds vector access in pekwm_sys TimeOfDay command.
 
 pekwm-0.4.1